test(solana): the boundary sweep becomes its own table-driven binary with its own snapshot
Extracts the ~950-line capacity instrument from host_mollusk.rs (now
behavior-only) into fhe_execute_boundary.rs: one shape table drives the
eleven sweeps, the per-shape WallPin asserts (builder-cap coverage, the
trace-one-under tie, and the invariant #61 host-heap gap with its
direction pinned so it cannot drift silently), the snapshot, and the
#[ignore]d full-curve matrix printer. Shared instruction fixtures move
to tests/host_fixtures/.

Distinct test binaries use distinct snapshot files, so the boundaries
and ceilings sections move from host_mollusk.json to a new
fhe_execute_boundary.json. Regenerated via update-cost-snapshots.sh:
all 11 boundaries and the host measurements are byte-identical; the
token/batcher app measurements rise 0.3-0.7% CU, the price of the
per-step build-heap gate in the previous commit.
