+++ solana/runtime-tests/cost-snapshots/token_mollusk.json
@@ -6,7 +6,7 @@
   },
   "measurements": {
     "confidential_burn_from_value/direct": {
-      "compute_units": 186429,
+      "compute_units": 187669,
       "unique_accounts": 16,
       "instruction_data_bytes": 8,
       "cpi_instructions": 11,
@@ -14,7 +14,7 @@
       "max_cpi_instruction_data_bytes": 665
     },
     "confidential_transfer/direct": {
-      "compute_units": 228691,
+      "compute_units": 229902,
       "unique_accounts": 14,
       "instruction_data_bytes": 223,
       "cpi_instructions": 7,
@@ -22,7 +22,7 @@
       "max_cpi_instruction_data_bytes": 1093
     },
     "confidential_transfer/steady_state": {
-      "compute_units": 236661,
+      "compute_units": 237871,
       "unique_accounts": 14,
       "instruction_data_bytes": 223,
       "cpi_instructions": 5,
@@ -30,7 +30,7 @@
       "max_cpi_instruction_data_bytes": 1225
     },
     "confidential_transfer_from_value/direct": {
-      "compute_units": 185432,
+      "compute_units": 186657,
       "unique_accounts": 15,
       "instruction_data_bytes": 8,
       "cpi_instructions": 7,
@@ -38,7 +38,7 @@
       "max_cpi_instruction_data_bytes": 699
     },
     "initialize_token_account": {
-      "compute_units": 56360,
+      "compute_units": 56713,
       "unique_accounts": 11,
       "instruction_data_bytes": 8,
       "cpi_instructions": 6,
