+++ solana/runtime-tests/tests/host_mollusk.rs
@@ -30,7 +30,7 @@ use solana_sdk::{
     instruction::{AccountMeta, Instruction},
     pubkey::Pubkey,
 };
-use std::collections::{BTreeMap, HashMap};
+use std::collections::HashMap;
 use zama_host::encode::ExecutionDictionary;
 use zama_host::{
     self as host, EncryptedValue, FheBinaryOpCode, FheExecuteArgs, FheExecuteOperand,
@@ -43,29 +43,19 @@ use zama_solana_test_kit::{
     host_svm_without_previous_bank_hash as mollusk_without_previous_bank_hash, label,
     new_encrypted_value as new_encrypted_value_account, read_encrypted_value,
     read_encrypted_value_from_result, readonly, readonly_signer, serialized_account, signing,
-    system_account, system_program_account, u256_be, writable, HostConfigParams,
-    DECRYPTION_CONTRACT, GATEWAY_CHAIN_ID, INPUT_VERIFICATION_CONTRACT,
+    system_account, system_program_account, writable, DECRYPTION_CONTRACT, GATEWAY_CHAIN_ID,
+    INPUT_VERIFICATION_CONTRACT,
+};
+
+mod host_fixtures;
+use host_fixtures::{
+    created_public_batch, fhe_execute_ix, fhe_execute_ix_with_deny_records, host_config_account,
+    host_config_account_with_flags,
 };
 
 // ---------------------------------------------------------------------------
 // Harness
 // ---------------------------------------------------------------------------
-
-fn host_config_account_with_flags(
-    admin: Pubkey,
-    paused: bool,
-    grant_deny_list_enabled: bool,
-) -> (Pubkey, Account) {
-    zama_solana_test_kit::host_config_account(&HostConfigParams {
-        paused,
-        grant_deny_list_enabled,
-        ..HostConfigParams::new(admin)
-    })
-}
-
-fn host_config_account(admin: Pubkey) -> (Pubkey, Account) {
-    host_config_account_with_flags(admin, false, false)
-}
 
 fn paused_host_config_account(admin: Pubkey) -> (Pubkey, Account) {
     host_config_account_with_flags(admin, true, false)
@@ -319,27 +309,6 @@ fn make_handle_public_ix_with_deny(
     )
 }
 
-/// Builds an `fhe_execute` instruction. `remaining` accounts are appended in
-/// order and referenced by index from `args`.
-fn fhe_execute_ix(
-    payer: Pubkey,
-    compute_subject: Pubkey,
-    encrypted_value_account_authority: Pubkey,
-    host_config: Pubkey,
-    args: FheExecuteArgs,
-    remaining: Vec<AccountMeta>,
-) -> Instruction {
-    fhe_execute_ix_with_deny_records(
-        payer,
-        compute_subject,
-        encrypted_value_account_authority,
-        host_config,
-        args,
-        remaining,
-        Vec::new(),
-    )
-}
-
 fn fhe_execute_ix_with_deny(
     payer: Pubkey,
     compute_subject: Pubkey,
@@ -358,43 +327,6 @@ fn fhe_execute_ix_with_deny(
         remaining,
         deny_subject_record.into_iter().collect(),
     )
-}
-
-fn fhe_execute_ix_with_deny_records(
-    payer: Pubkey,
-    compute_subject: Pubkey,
-    encrypted_value_account_authority: Pubkey,
-    host_config: Pubkey,
-    mut args: FheExecuteArgs,
-    remaining: Vec<AccountMeta>,
-    deny_subject_records: Vec<Pubkey>,
-) -> Instruction {
-    // The execution self-describes its `remaining_accounts` length (DD-033); fixtures
-    // build the account list here, so the declared count is stamped here too.
-    args.account_count = u8::try_from(remaining.len() + deny_subject_records.len())
-        .expect("fixture remaining accounts fit u8");
-    let mut ix = anchor_ix(
-        host::id(),
-        host::accounts::FheExecute {
-            payer,
-            compute_subject,
-            encrypted_value_account_authority,
-            host_config,
-            system_program: system_program::ID,
-            // Unrestricted block cap (u64::MAX) in every existing fixture: block_cap
-            // short-circuits before touching the optional accounts, so the two HCU
-            // witnesses stay absent.
-            hcu_block_meter: None,
-            hcu_trusted_app_record: None,
-            event_authority: event_authority(host::id()),
-            program: host::id(),
-        },
-        host::instruction::FheExecute { args },
-    );
-    ix.accounts.extend(remaining);
-    ix.accounts
-        .extend(deny_subject_records.into_iter().map(readonly));
-    ix
 }
 
 #[test]
@@ -2317,100 +2249,6 @@ fn mollusk_fhe_execute_updates_persistent_output_with_previous_state() {
 // fhe_execute: narrow produced-public lifecycle execution
 // ---------------------------------------------------------------------------
 
-struct CreatedPublicBatch {
-    instruction: Instruction,
-    accounts: Vec<(Pubkey, Account)>,
-    outputs: Vec<(u16, Pubkey)>,
-}
-
-fn created_public_batch(step_count: usize, created_public_steps: &[usize]) -> CreatedPublicBatch {
-    let authority = Pubkey::new_unique();
-    persistent_creates_batch(
-        step_count,
-        created_public_steps,
-        authority,
-        true,
-        std::slice::from_ref(&authority),
-    )
-}
-
-/// [`created_public_batch`] with a caller-fixed authority (for boundary sweeps recorded in the
-/// cost snapshot: PDA bump searches are part of measured compute, so recorded profiles need
-/// stable keys) and a caller-chosen `make_public` — `false` gives the plain persistent create,
-/// the shape `zama-fhe`'s `heap_budget.rs` measures on the app side.
-fn persistent_creates_batch(
-    step_count: usize,
-    created_public_steps: &[usize],
-    authority: Pubkey,
-    make_public: bool,
-    subjects: &[Pubkey],
-) -> CreatedPublicBatch {
-    let (host_config, host_config_account) = host_config_account(authority);
-    let mut output_metas = Vec::new();
-    let mut output_accounts = Vec::new();
-    let mut outputs = Vec::new();
-    let mut steps = Vec::with_capacity(step_count);
-    let mut dictionary = ExecutionDictionary::default();
-
-    for step_index in 0..step_count {
-        let output = if created_public_steps.contains(&step_index) {
-            let output_label = label(&format!("created-public-{step_index}"));
-            let encrypted_value_id = zama_solana_acl::derive_encrypted_value_id(
-                authority.to_bytes(),
-                authority.to_bytes(),
-                output_label,
-            );
-            let output_address = host::encrypted_value_address(encrypted_value_id).0;
-            let output_index = u8::try_from(output_metas.len()).unwrap();
-            output_metas.push(writable(output_address));
-            output_accounts.push((output_address, empty_system_account()));
-            outputs.push((step_index as u16, output_address));
-            FheExecuteOutput::StoredValue {
-                output_encrypted_value_index: output_index,
-                output_authority_index: None,
-                output_domain_index: dictionary.intern_key(authority),
-                output_account_index: dictionary.intern_key(authority),
-                output_label_index: dictionary.intern(output_label),
-                output_subject_indexes: dictionary.intern_subjects(subjects.iter().copied()),
-                previous_state: None,
-                make_public,
-            }
-        } else {
-            FheExecuteOutput::Transient
-        };
-        steps.push(FheExecuteStep::TrivialEncrypt {
-            plaintext: [(step_index + 1) as u8; 32],
-            fhe_type: 5,
-            output,
-        });
-    }
-
-    let instruction = fhe_execute_ix(
-        authority,
-        authority,
-        authority,
-        host_config,
-        FheExecuteArgs {
-            account_count: 0,
-            dictionary: dictionary.into_entries(),
-            steps,
-        },
-        output_metas,
-    );
-    let mut accounts = vec![
-        (system_program::ID, system_program_account()),
-        (authority, funded_system_account()),
-        (host_config, host_config_account),
-        (event_authority(host::id()), Account::default()),
-    ];
-    accounts.extend(output_accounts);
-    CreatedPublicBatch {
-        instruction,
-        accounts,
-        outputs,
-    }
-}
-
 /// The bytes an event-CPI inner instruction starts with: Anchor's event-CPI instruction tag, then
 /// the event's own discriminator.
 fn event_cpi_prefix<T: Discriminator>() -> Vec<u8> {
@@ -2592,789 +2430,6 @@ fn mollusk_fhe_execute_created_public_heap_boundary() {
 // largest passing run. `cost_snapshot::assert_boundary_snapshot` pins all of
 // it, so a boundary that moves (or changes walls) arrives as a reviewed diff.
 // ===========================================================================
-
-/// Names the wall a failing probe hit, so a boundary sweep can never silently converge on the
-/// wrong limit. `heap` is the SBF abort (`ProgramFailedToComplete`), which also covers panics:
-/// Mollusk's `InstructionResult` carries no program logs, so the allocator's "memory allocation
-/// failed" line cannot be read back here to separate them. A probe asserting `heap` must
-/// therefore use a shape known to be panic-free at smaller sizes — which the sweep's monotonic
-/// prefix requirement enforces in practice, since a panicking fixture fails at every size, not
-/// past a boundary. Program errors (`other:*`) are never a capacity wall; `sweep_step_boundary`
-/// panics on them instead of recording them.
-fn failure_axis(result: &mollusk_svm::result::InstructionResult) -> String {
-    use solana_sdk::instruction::InstructionError;
-    match &result.raw_result {
-        Ok(()) => "none".to_string(),
-        Err(InstructionError::ProgramFailedToComplete) => "heap".to_string(),
-        Err(InstructionError::ComputationalBudgetExceeded) => "compute_units".to_string(),
-        Err(InstructionError::MaxInstructionTraceLengthExceeded) => "instruction_trace".to_string(),
-        Err(other) => format!("other:{other:?}"),
-    }
-}
-
-/// One probe of a boundary sweep: a complete instruction plus the accounts it runs against.
-struct ProbeCase {
-    instruction: Instruction,
-    accounts: Vec<(Pubkey, Account)>,
-}
-
-impl From<CreatedPublicBatch> for ProbeCase {
-    fn from(case: CreatedPublicBatch) -> Self {
-        ProbeCase {
-            instruction: case.instruction,
-            accounts: case.accounts,
-        }
-    }
-}
-
-/// The result of sweeping one shape: the largest passing step count, the first failing one, the
-/// wall that failure hit, and the passing run's evidence for the snapshot.
-struct SweptBoundary {
-    max_ok: usize,
-    first_fail: usize,
-    limited_by: String,
-    instruction_at_max_ok: Instruction,
-    result_at_max_ok: mollusk_svm::result::InstructionResult,
-}
-
-/// Sweeps a shape from `min_steps` to the host's step cap. The sweep is linear rather than a
-/// binary search so it can assert the boundary premise itself: success must be a prefix of the
-/// range. A success after a failure would mean step count is not what governs this shape, and a
-/// recorded boundary would be meaningless.
-fn sweep_step_boundary(min_steps: usize, build: impl Fn(usize) -> ProbeCase) -> SweptBoundary {
-    let mut last_ok: Option<(usize, Instruction, mollusk_svm::result::InstructionResult)> = None;
-    let mut first_failure: Option<(usize, String)> = None;
-    for steps in min_steps..=host::MAX_FHE_EXECUTION_STEPS {
-        let case = build(steps);
-        let result = mollusk().process_instruction(&case.instruction, &case.accounts);
-        let axis = failure_axis(&result);
-        assert!(
-            !axis.starts_with("other:"),
-            "probe at {steps} steps failed on a program error, not a capacity wall: {axis} — \
-             fix the fixture instead of recording a boundary"
-        );
-        if axis == "none" {
-            if let Some((failed_steps, failed_axis)) = &first_failure {
-                panic!(
-                    "non-monotonic boundary: {steps} steps succeeded after {failed_steps} \
-                     failed on {failed_axis}; step count does not govern this shape"
-                );
-            }
-            last_ok = Some((steps, case.instruction, result));
-        } else if first_failure.is_none() {
-            first_failure = Some((steps, axis));
-        }
-    }
-    let (max_ok, instruction_at_max_ok, result_at_max_ok) =
-        last_ok.expect("the smallest probe of the sweep must pass");
-    let (first_fail, limited_by) = first_failure.unwrap_or((
-        // Every legal size passes: the policy cap is the wall, and no runtime wall was
-        // observed below it.
-        host::MAX_FHE_EXECUTION_STEPS + 1,
-        "MAX_FHE_EXECUTION_STEPS".to_string(),
-    ));
-    SweptBoundary {
-        max_ok,
-        first_fail,
-        limited_by,
-        instruction_at_max_ok,
-        result_at_max_ok,
-    }
-}
-
-fn assert_swept_boundary(profile: &str, sweep: &SweptBoundary) {
-    cost_snapshot::assert_boundary_snapshot(
-        "host_mollusk",
-        profile,
-        &cost_snapshot::Boundary {
-            max_ok: sweep.max_ok as u64,
-            first_fail: sweep.first_fail as u64,
-            limited_by: sweep.limited_by.clone(),
-        },
-        &sweep.instruction_at_max_ok,
-        &sweep.result_at_max_ok,
-    );
-}
-
-/// A dependent transient chain closing in one persistent create: every step adds a scalar to the
-/// previous step's transient result. This is the transient-heavy shape `dep-chain` and the
-/// load-smoke scenario drive, and the lightest per-step shape in the matrix.
-fn dependent_chain_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    assert!(
-        steps >= 2,
-        "the chain shape needs a first read and a final persist"
-    );
-    let (host_config, host_config_account) = host_config_account(authority);
-    let balance_handle = handle_for_chain(0x61, 5);
-    let (balance_address, balance_value) = new_encrypted_value_account(
-        authority,
-        authority,
-        label("boundary-chain-balance"),
-        balance_handle,
-        &[authority],
-    );
-    let output_label = label("boundary-chain-output");
-    let output_id = zama_solana_acl::derive_encrypted_value_id(
-        authority.to_bytes(),
-        authority.to_bytes(),
-        output_label,
-    );
-    let output_address = host::encrypted_value_address(output_id).0;
-
-    let mut dictionary = ExecutionDictionary::default();
-    let one = dictionary.intern(u256_be(1));
-    let balance_handle_index = dictionary.intern(balance_handle);
-    let mut chain = Vec::with_capacity(steps);
-    chain.push(FheExecuteStep::Binary {
-        op: FheBinaryOpCode::Add,
-        lhs: FheExecuteOperand::StoredValue {
-            handle_index: balance_handle_index,
-            encrypted_value_index: 0,
-        },
-        rhs: FheExecuteOperand::Scalar { value_index: one },
-        output_fhe_type: 5,
-        output: FheExecuteOutput::Transient,
-    });
-    for index in 1..steps {
-        let output = if index == steps - 1 {
-            FheExecuteOutput::StoredValue {
-                output_encrypted_value_index: 1,
-                output_authority_index: None,
-                output_domain_index: dictionary.intern_key(authority),
-                output_account_index: dictionary.intern_key(authority),
-                output_label_index: dictionary.intern(output_label),
-                output_subject_indexes: dictionary.intern_subjects([authority]),
-                previous_state: None,
-                make_public: false,
-            }
-        } else {
-            FheExecuteOutput::Transient
-        };
-        chain.push(FheExecuteStep::Binary {
-            op: FheBinaryOpCode::Add,
-            lhs: FheExecuteOperand::EarlierStep {
-                producer_index: (index - 1) as u8,
-            },
-            rhs: FheExecuteOperand::Scalar { value_index: one },
-            output_fhe_type: 5,
-            output,
-        });
-    }
-    let instruction = fhe_execute_ix(
-        authority,
-        authority,
-        authority,
-        host_config,
-        FheExecuteArgs {
-            account_count: 0,
-            dictionary: dictionary.into_entries(),
-            steps: chain,
-        },
-        vec![readonly(balance_address), writable(output_address)],
-    );
-    ProbeCase {
-        instruction,
-        accounts: vec![
-            (system_program::ID, system_program_account()),
-            (authority, funded_system_account()),
-            (host_config, host_config_account),
-            (event_authority(host::id()), Account::default()),
-            (balance_address, encrypted_value_account(&balance_value)),
-            (output_address, empty_system_account()),
-        ],
-    }
-}
-
-/// Every step updates its own mature `EncryptedValue` carrying `peak_count` MMR peaks and the
-/// full subject list, so each account decode allocates what the given maturity forces. The leaf
-/// count keeps eight trailing zero bits so the eight seal-appends per account stay cheap — the
-/// shape stresses decode allocation, not MMR merge hashing. Peak count is `popcount(leaf_count)`
-/// — pure on-chain state the builder cannot see, which is why this axis is swept per maturity
-/// instead of enforced at build time.
-fn mature_updates_case(steps: usize, peak_count: u32, authority: Pubkey) -> ProbeCase {
-    let (host_config, host_config_account) = host_config_account(authority);
-    let subjects: Vec<Pubkey> = (1..=zama_solana_acl::MAX_ENCRYPTED_VALUE_SUBJECTS as u8)
-        .map(|index| Pubkey::new_from_array([0x70 + index; 32]))
-        .collect();
-    // `peak_count` one-bits, 8 trailing zeros (cheap appends).
-    assert!(
-        (1..=55).contains(&peak_count),
-        "peak_count must keep 8 trailing zeros free"
-    );
-    let leaf_count: u64 = (((1u128 << peak_count) - 1) as u64) << 8;
-    let peaks: Vec<[u8; 32]> = (0..leaf_count.count_ones())
-        .map(|peak| [peak as u8 + 1; 32])
-        .collect();
-
-    let mut dictionary = ExecutionDictionary::default();
-    let domain_index = dictionary.intern_key(authority);
-    let account_index = dictionary.intern_key(authority);
-    let subject_indexes = dictionary.intern_subjects(subjects.iter().copied());
-    let mut update_steps = Vec::with_capacity(steps);
-    let mut metas = Vec::with_capacity(steps);
-    let mut accounts = vec![
-        (system_program::ID, system_program_account()),
-        (authority, funded_system_account()),
-        (host_config, host_config_account),
-        (event_authority(host::id()), Account::default()),
-    ];
-    for step_index in 0..steps {
-        let value_label = label(&format!("boundary-mature-{step_index}"));
-        let handle = handle_for_chain(0x80 + step_index as u8, 5);
-        let (address, mut value) =
-            new_encrypted_value_account(authority, authority, value_label, handle, &subjects);
-        value.leaf_count = leaf_count;
-        value.peaks = peaks.clone();
-        metas.push(writable(address));
-        accounts.push((address, encrypted_value_account(&value)));
-        update_steps.push(FheExecuteStep::TrivialEncrypt {
-            plaintext: [step_index as u8 + 1; 32],
-            fhe_type: 5,
-            output: FheExecuteOutput::StoredValue {
-                output_encrypted_value_index: step_index as u8,
-                output_authority_index: None,
-                output_domain_index: domain_index,
-                output_account_index: account_index,
-                output_label_index: dictionary.intern(value_label),
-                output_subject_indexes: subject_indexes.clone(),
-                previous_state: Some(PreviousState {
-                    handle,
-                    subjects: subjects.clone(),
-                }),
-                make_public: false,
-            },
-        });
-    }
-    let instruction = fhe_execute_ix(
-        authority,
-        authority,
-        authority,
-        host_config,
-        FheExecuteArgs {
-            account_count: 0,
-            dictionary: dictionary.into_entries(),
-            steps: update_steps,
-        },
-        metas,
-    );
-    ProbeCase {
-        instruction,
-        accounts,
-    }
-}
-
-/// Every step consumes its own maximum-size verified input: an inline attestation covering
-/// `MAX_INPUT_ATTESTATION_HANDLES` handles with `MAX_INPUT_ATTESTATION_EXTRA_DATA` bytes of extra
-/// data, signed at the fixture threshold of one. Each attestation is decoded and re-verified
-/// (one secp256k1 recovery per signature) in-execution.
-fn attestation_per_step_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let signer_key = signing::coprocessor_signing_key();
-    let (host_config, host_config_account) =
-        zama_solana_test_kit::host_config_account(&HostConfigParams {
-            coprocessor_signers: vec![signing::secp_evm_address(&signer_key)],
-            coprocessor_threshold: 1,
-            ..HostConfigParams::new(authority)
-        });
-    let mut dictionary = ExecutionDictionary::default();
-    let one = dictionary.intern(u256_be(1));
-    let attestation_steps = (0..steps)
-        .map(|step_index| {
-            let ct_handles: Vec<[u8; 32]> = (0..host::MAX_INPUT_ATTESTATION_HANDLES)
-                .map(|position| {
-                    let mut handle = handle_for_chain(0x90 + step_index as u8, 5);
-                    // The host requires each covered handle to carry its own position.
-                    handle[21] = position as u8;
-                    handle
-                })
-                .collect();
-            let attestation = signing::attestation_signed_by(
-                ct_handles[0],
-                ct_handles,
-                0,
-                authority,
-                authority,
-                vec![0xAB; host::MAX_INPUT_ATTESTATION_EXTRA_DATA],
-                std::slice::from_ref(&signer_key),
-            );
-            FheExecuteStep::Binary {
-                op: FheBinaryOpCode::Add,
-                lhs: FheExecuteOperand::VerifiedInput {
-                    attestation: Box::new(attestation),
-                },
-                rhs: FheExecuteOperand::Scalar { value_index: one },
-                output_fhe_type: 5,
-                output: FheExecuteOutput::Transient,
-            }
-        })
-        .collect();
-    let instruction = fhe_execute_ix(
-        authority,
-        authority,
-        authority,
-        host_config,
-        FheExecuteArgs {
-            account_count: 0,
-            dictionary: dictionary.into_entries(),
-            steps: attestation_steps,
-        },
-        Vec::new(),
-    );
-    ProbeCase {
-        instruction,
-        accounts: vec![
-            (system_program::ID, system_program_account()),
-            (authority, funded_system_account()),
-            (host_config, host_config_account),
-            (event_authority(host::id()), Account::default()),
-        ],
-    }
-}
-
-fn all_created_public_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let all: Vec<usize> = (0..steps).collect();
-    persistent_creates_batch(
-        steps,
-        &all,
-        authority,
-        true,
-        std::slice::from_ref(&authority),
-    )
-    .into()
-}
-
-/// Every step writes a plain persistent create (no `make_public`) — the exact shape
-/// `zama-fhe`'s `heap_budget.rs` measures on the app side, so its host-side wall is on record
-/// next to the app-side byte count that motivates the single step ceiling.
-fn all_private_creates_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let all: Vec<usize> = (0..steps).collect();
-    persistent_creates_batch(
-        steps,
-        &all,
-        authority,
-        false,
-        std::slice::from_ref(&authority),
-    )
-    .into()
-}
-
-/// Every fourth step persists a created-public output, the rest stay transient — a mid-weight
-/// composite between the chain and all-created-public extremes.
-fn mixed_chain_creates_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let creates: Vec<usize> = (0..steps).step_by(4).collect();
-    persistent_creates_batch(
-        steps,
-        &creates,
-        authority,
-        true,
-        std::slice::from_ref(&authority),
-    )
-    .into()
-}
-
-/// Every step creates a persistent output with the widest legal audience: eight distinct
-/// subjects, each interning its own dictionary entry and each resolved and stored by the host.
-/// The subject-width axis of the create frontier — the builder's build-heap budget stops this
-/// shape well below the host's own wall.
-fn subject_heavy_creates_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let all: Vec<usize> = (0..steps).collect();
-    let subjects: Vec<Pubkey> = (1..=zama_solana_acl::MAX_ENCRYPTED_VALUE_SUBJECTS as u8)
-        .map(|index| Pubkey::new_from_array([0x50 + index; 32]))
-        .collect();
-    persistent_creates_batch(steps, &all, authority, false, &subjects).into()
-}
-
-/// [`subject_heavy_creates_case`] with every output also made public — the corner where the
-/// widest audience meets the public-outputs event. The one-subject public sweep sits exactly on
-/// the trace-heap boundary with zero margin, so this axis pair is measured rather than
-/// interpolated from the two single-axis sweeps.
-fn subject_heavy_public_creates_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let all: Vec<usize> = (0..steps).collect();
-    let subjects: Vec<Pubkey> = (1..=zama_solana_acl::MAX_ENCRYPTED_VALUE_SUBJECTS as u8)
-        .map(|index| Pubkey::new_from_array([0x60 + index; 32]))
-        .collect();
-    persistent_creates_batch(steps, &all, authority, true, &subjects).into()
-}
-
-/// Every step past the first is a `Sum` over the coprocessor's maximum euint64 operand count,
-/// all referencing the previous step's transient (distinct operands keep the derived handles
-/// distinct) — the operand-width axis: per step the host allocates its operand tables
-/// proportionally and meters every operand's HCU.
-fn reduction_heavy_case(steps: usize, authority: Pubkey) -> ProbeCase {
-    let (host_config, host_config_account) = host_config_account(authority);
-    let mut steps_vec = Vec::with_capacity(steps);
-    steps_vec.push(FheExecuteStep::TrivialEncrypt {
-        plaintext: [1; 32],
-        fhe_type: 5,
-        output: FheExecuteOutput::Transient,
-    });
-    for step_index in 1..steps {
-        steps_vec.push(FheExecuteStep::Sum {
-            operands: vec![
-                FheExecuteOperand::EarlierStep {
-                    producer_index: (step_index - 1) as u8
-                };
-                60
-            ],
-            fhe_type: 5,
-            output: FheExecuteOutput::Transient,
-        });
-    }
-    let instruction = fhe_execute_ix(
-        authority,
-        authority,
-        authority,
-        host_config,
-        FheExecuteArgs {
-            account_count: 0,
-            dictionary: ExecutionDictionary::default().into_entries(),
-            steps: steps_vec,
-        },
-        Vec::new(),
-    );
-    ProbeCase {
-        instruction,
-        accounts: vec![
-            (system_program::ID, system_program_account()),
-            (authority, funded_system_account()),
-            (host_config, host_config_account),
-            (event_authority(host::id()), Account::default()),
-        ],
-    }
-}
-
-/// The largest create count `zama-fhe`'s instruction-trace model admits: the greatest count
-/// whose floor — including the app wrapper instruction the builder budgets for — still fits the
-/// transaction's trace.
-fn max_creates_the_builder_admits(make_public: bool) -> usize {
-    (0..=host::MAX_FHE_EXECUTION_STEPS)
-        .take_while(|creates| {
-            zama_fhe::instruction_trace_floor(*creates, false, make_public)
-                <= zama_fhe::TRANSACTION_INSTRUCTION_TRACE_LIMIT
-        })
-        .last()
-        .expect("zero creates always fit")
-}
-
-#[test]
-fn cost_snapshot_boundary_all_created_public() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        all_created_public_case(steps, Pubkey::new_from_array([0x31; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/all_created_public", &sweep);
-    // The builder never admits a create count the host measured as failing. For this shape the
-    // heap and trace walls sit within one step of each other, so the builder's trace cap covers
-    // both.
-    assert!(
-        max_creates_the_builder_admits(true) <= sweep.max_ok,
-        "the builder admits {} created-public outputs but the host wall is {}",
-        max_creates_the_builder_admits(true),
-        sweep.max_ok,
-    );
-}
-
-#[test]
-fn cost_snapshot_boundary_all_private_creates() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        all_private_creates_case(steps, Pubkey::new_from_array([0x36; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/all_private_creates", &sweep);
-    // The tie between the builder's model and the measured wall: Mollusk probes the host
-    // top-level, one instruction short of the app wrapper the builder's floor budgets, so the
-    // host must run exactly one create past the builder's cap and no further.
-    assert_eq!(sweep.limited_by, "instruction_trace");
-    assert_eq!(
-        max_creates_the_builder_admits(false) + 1,
-        sweep.max_ok,
-        "the measured trace wall no longer sits one wrapper instruction past the builder's cap",
-    );
-}
-
-#[test]
-fn cost_snapshot_boundary_subject_heavy_creates() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        subject_heavy_creates_case(steps, Pubkey::new_from_array([0x37; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/subject_heavy_creates", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_subject_heavy_public_creates() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        subject_heavy_public_creates_case(steps, Pubkey::new_from_array([0x3B; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/subject_heavy_public_creates", &sweep);
-    // On this axis pair the wall is the HOST's own CPI frame — the created accounts' subject
-    // tables and the public-outputs event payload grow with creates x subjects-per-output —
-    // and it sits well below the trace cap, which the one-subject sweeps could not see. The
-    // builder models only the app's CPI frame (build + packet + invoke tables), and no
-    // app-side number can see this driver: with a shared audience the dictionary interns
-    // eight subjects once, so the builder admits the trace-capped twenty such creates while
-    // the host survives fifteen. Until the host's side gets its own typed ceiling (or a
-    // policy cap on the public payload), this wall is measured, not typed: the snapshot pins
-    // it, and the boundary matrix is the guidance — fhevm-internal#1872.
-    assert_eq!(
-        sweep.limited_by, "heap",
-        "the wide-audience public wall moved off the host heap — re-read the guidance above"
-    );
-}
-
-#[test]
-fn cost_snapshot_boundary_reduction_heavy() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        reduction_heavy_case(steps, Pubkey::new_from_array([0x38; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/reduction_heavy", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_dependent_chain() {
-    let sweep = sweep_step_boundary(2, |steps| {
-        dependent_chain_case(steps, Pubkey::new_from_array([0x32; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/dependent_chain", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_mature_updates() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        mature_updates_case(steps, 55, Pubkey::new_from_array([0x33; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/mature_updates", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_mature_updates_peaks_8() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        mature_updates_case(steps, 8, Pubkey::new_from_array([0x39; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/mature_updates_peaks_8", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_mature_updates_peaks_32() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        mature_updates_case(steps, 32, Pubkey::new_from_array([0x3A; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/mature_updates_peaks_32", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_attestation_per_step() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        attestation_per_step_case(steps, Pubkey::new_from_array([0x34; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/attestation_per_step", &sweep);
-}
-
-#[test]
-fn cost_snapshot_boundary_mixed_chain_creates() {
-    let sweep = sweep_step_boundary(1, |steps| {
-        mixed_chain_creates_case(steps, Pubkey::new_from_array([0x35; 32]))
-    });
-    assert_swept_boundary("fhe_execute_boundary/mixed_chain_creates", &sweep);
-}
-
-/// The limits every measurement in this file runs against, as data: what each is, whether a
-/// transaction can push it, and how. Solana values are runtime facts of the pinned toolchain;
-/// zama values are this repo's policy, asserted from the constants so the snapshot cannot drift
-/// from the code.
-#[test]
-fn cost_snapshot_solana_ceilings() {
-    let ceiling = |value: u64, extendable: bool, note: &str| cost_snapshot::Ceiling {
-        value,
-        extendable,
-        note: note.to_string(),
-    };
-    let ceilings = BTreeMap::from([
-        (
-            "solana/heap_frame_bytes".to_string(),
-            ceiling(
-                32 * 1024,
-                false,
-                "fixed-length bump allocator installed by solana-program-entrypoint; every \
-                 invocation (each top-level instruction and each CPI frame) gets a fresh region. \
-                 RequestHeapFrame grants up to 256 KiB per frame, but the default allocator's \
-                 length is a compile-time constant, so the grant is unusable without a custom \
-                 allocator (fhevm-internal#1872: we do not ship one)",
-            ),
-        ),
-        (
-            "solana/compute_units_per_instruction_default".to_string(),
-            ceiling(
-                200_000,
-                true,
-                "SetComputeUnitLimit raises it toward the transaction cap; the priority fee is \
-                 charged on the requested limit, not on usage",
-            ),
-        ),
-        (
-            "solana/compute_units_per_transaction_max".to_string(),
-            ceiling(
-                1_400_000,
-                false,
-                "hard runtime cap; no instruction raises it",
-            ),
-        ),
-        (
-            "solana/transaction_bytes".to_string(),
-            ceiling(
-                1_232,
-                false,
-                "IPv6-MTU packet limit; address lookup tables compress account keys, not \
-                 instruction data",
-            ),
-        ),
-        (
-            "solana/instruction_trace_length".to_string(),
-            ceiling(
-                64,
-                false,
-                "every instruction executed in a transaction, top-level plus each CPI; sits \
-                 within one step of the all-created-public shape's heap wall (each created \
-                 output issues ~3 CPIs), and it is shared with the app's own CPIs in the same \
-                 transaction",
-            ),
-        ),
-        (
-            "solana/cpi_instruction_data_bytes".to_string(),
-            ceiling(10_240, false, "per CPI call"),
-        ),
-        (
-            "solana/cpi_call_depth".to_string(),
-            ceiling(5, false, "maximum invoke nesting"),
-        ),
-        (
-            "solana/stack_frame_bytes".to_string(),
-            ceiling(4_096, false, "per stack frame"),
-        ),
-        (
-            "solana/return_data_bytes".to_string(),
-            ceiling(1_024, false, "per set_return_data"),
-        ),
-        (
-            "zama/max_fhe_execution_steps".to_string(),
-            ceiling(
-                host::MAX_FHE_EXECUTION_STEPS as u64,
-                true,
-                "policy cap owned by this repo; raising it is a code change plus a re-run of \
-                 the boundary sweeps above (fhevm-internal#1872)",
-            ),
-        ),
-        (
-            "zama/max_input_attestation_handles".to_string(),
-            ceiling(
-                host::MAX_INPUT_ATTESTATION_HANDLES as u64,
-                true,
-                "policy: covered handles per inline attestation",
-            ),
-        ),
-        (
-            "zama/max_input_attestation_extra_data_bytes".to_string(),
-            ceiling(
-                host::MAX_INPUT_ATTESTATION_EXTRA_DATA as u64,
-                true,
-                "policy: extra-data bytes per inline attestation",
-            ),
-        ),
-        (
-            "zama/max_mmr_peaks".to_string(),
-            ceiling(
-                zama_solana_acl::MAX_MMR_PEAKS as u64,
-                false,
-                "structural: a u64 leaf count can never hold more than 64 peaks",
-            ),
-        ),
-        (
-            "zama/max_encrypted_value_subjects".to_string(),
-            ceiling(
-                zama_solana_acl::MAX_ENCRYPTED_VALUE_SUBJECTS as u64,
-                true,
-                "policy: subjects per encrypted value",
-            ),
-        ),
-    ]);
-    cost_snapshot::assert_ceilings_snapshot("host_mollusk", &ceilings);
-}
-
-/// One row of the printed matrix: shape name, its minimum legal step count, and its case builder.
-type ProbeShape = (&'static str, usize, Box<dyn Fn(usize) -> ProbeCase>);
-
-#[test]
-#[ignore = "full boundary curves, run explicitly with --nocapture"]
-fn print_boundary_matrix() {
-    let shapes: Vec<ProbeShape> = vec![
-        (
-            "all_created_public",
-            1,
-            Box::new(|steps| all_created_public_case(steps, Pubkey::new_from_array([0x31; 32]))),
-        ),
-        (
-            "all_private_creates",
-            1,
-            Box::new(|steps| all_private_creates_case(steps, Pubkey::new_from_array([0x36; 32]))),
-        ),
-        (
-            "subject_heavy_public_creates",
-            1,
-            Box::new(|steps| {
-                subject_heavy_public_creates_case(steps, Pubkey::new_from_array([0x3B; 32]))
-            }),
-        ),
-        (
-            "dependent_chain",
-            2,
-            Box::new(|steps| dependent_chain_case(steps, Pubkey::new_from_array([0x32; 32]))),
-        ),
-        (
-            "mature_updates",
-            1,
-            Box::new(|steps| mature_updates_case(steps, 55, Pubkey::new_from_array([0x33; 32]))),
-        ),
-        (
-            "mature_updates_peaks_8",
-            1,
-            Box::new(|steps| mature_updates_case(steps, 8, Pubkey::new_from_array([0x39; 32]))),
-        ),
-        (
-            "mature_updates_peaks_32",
-            1,
-            Box::new(|steps| mature_updates_case(steps, 32, Pubkey::new_from_array([0x3A; 32]))),
-        ),
-        (
-            "attestation_per_step",
-            1,
-            Box::new(|steps| attestation_per_step_case(steps, Pubkey::new_from_array([0x34; 32]))),
-        ),
-        (
-            "mixed_chain_creates",
-            1,
-            Box::new(|steps| mixed_chain_creates_case(steps, Pubkey::new_from_array([0x35; 32]))),
-        ),
-        (
-            "subject_heavy_creates",
-            1,
-            Box::new(|steps| subject_heavy_creates_case(steps, Pubkey::new_from_array([0x37; 32]))),
-        ),
-        (
-            "reduction_heavy",
-            1,
-            Box::new(|steps| reduction_heavy_case(steps, Pubkey::new_from_array([0x38; 32]))),
-        ),
-    ];
-    for (name, min_steps, build) in shapes {
-        for steps in min_steps..=host::MAX_FHE_EXECUTION_STEPS {
-            let case = build(steps);
-            let result = mollusk().process_instruction(&case.instruction, &case.accounts);
-            println!(
-                "{name:22} steps={steps:2} cu={:7} ix_bytes={:5} cpis={:2} axis={}",
-                result.compute_units_consumed,
-                case.instruction.data.len(),
-                result.inner_instructions.len(),
-                failure_axis(&result),
-            );
-        }
-    }
-}
 
 #[test]
 fn mollusk_fhe_execute_wrong_event_authority_fails_without_output() {
